Trying to fix the numbered list.
Currently the numbered lists restart their numbering after they hit the embedded code block. However, this is only an issue on the actual website (it works fine for the GitHub preview page). Here's an attempt to fix it: increasing the indenting for the code block. This may or not have an effect since your markdown to HTML conversion is likely different than GitHub's.